requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0
flasgger==0.9.7.1
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
redis==4.6.0
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
numpy>=1.24.0
pytesseract==0.3.13
Pillow>=10.0.0
opencv-python>=4.8.0
pdf2image>=1.16.0
openai>=1.0.0
logging==0.4.9.6
typing-extensions>=4.9.0
selenium==4.18.1
webdriver-manager==4.0.1
pymongo==4.6.1
motor==3.3.2
fastapi==0.110.0
uvicorn==0.27.1
tqdm==4.66.2
